refactor: Character classes file structure
