+++ Jogo Luta/src/js/characters/Knight.js
@@ -1,0 +1,12 @@
+import { Character } from "./Character.js";
+
+export class Knight extends Character {
+    constructor(name) {
+        super(name);
+        this.life = 100;
+        this.attack = 10;
+        this.defense = 8;
+        this.maxLife = this.life;
+        this.alive = this.alive;
+    }
+}
