+++ Jogo Luta/src/js/characters/Sorcerer.js
@@ -1,0 +1,12 @@
+import { Character } from "./Character.js";
+
+export class Sorcerer extends Character {
+    constructor(name) {
+        super(name);
+        this.life = 80;
+        this.attack = 15;
+        this.defense = 3;
+        this.maxLife = this.life;
+        this.alive = this.alive;
+    }
+}
