+++ Jogo Luta/src/js/characters/index.js
@@ -1,0 +1,4 @@
+export { Knight } from "./Knight.js";
+export { Sorcerer } from "./Sorcerer.js";
+export { LittleMonster } from "./LittleMonster.js";
+export { BigMonster } from "./BigMonster.js";
